Update plans should reference IDs for existing events
The CSV representation of the calendar has no concept of identity, so when
working out what to add/remove we compare the event data. A change in any of
the fields means that the old event will be removed and readded with the same
data.

However, when interacting with the remote calendar, we need to use the event IDs
to actually delete stuff, so the "to_remove" set should be a set of IDs.